Bug Fixes And Quality of Life Improvements
* Fixed issues with genome indexing & download
* Moved all functions to functions.sh and renamed the script
* Somehow the speed of samtools faidx is faster when using named pipes

+++ extract_genomic_regions.sh
@@ -9,82 +9,6 @@
 # $3 = input gene list
 # $4 = org name(eg x_laevis)
 
-##FUNCTIONS
-
-function get_fasta() {
-#1 - gene name
-#2 - transcript region (cds/5utr/3utr/exons/noncodingexons)
-#3 - basename for bedfiles
-#4 - genome fasta file
-#5 - org name
-#6 - FASTA OUPUT FOLDER
-#7 - Formatted file name for gene (without punctuation)
-#8 - GTF file path (for extracting the 'real' gene annotation)
-#9 - TEMP PATH
-
-local gene_full=$1
-local s_name=$(echo $gene_full | awk '{ gsub(/[[:punct:]]/, "_", $0) } 1;')
-local reg=$2
-local base_bed=$3
-local genome_fa=$4
-local org_name=$5
-local FASTA_PATH=$6
-local GTF_PATH=$7
-local TEMP_PATH=$8
-local LABEL_FASTA=$9
-
-if [ ! -s $FASTA_PATH/$s_name.$reg ]; then #$FASTA_PATH/$file_out.cds
-  >&2 echo $s_name $reg
-  
-  grep -i -w $gene_full files/genes/$org_name/gtf_stats.csv | awk -F',' '{print $3}' | grep -w -f - "$base_bed"_"$reg.bed" > $TEMP_PATH/"$s_name"_"$reg.bed"
-
-  ##TO get flanks 
-  #grep -i -f files/genes/some_org/cat1.rna_list files/genes/some_org/cat1.gtf_slice | grep -i "utr" | grep -i "three\|3"
-
-  ##TO find MEDIAN values
-  #grep -i -f files/genes/some_org/cat1.rna_list files/genes/some_org/cat1.gtf_slice | grep -i "utr" | grep -i "three\|3" | awk '{print $5-$4}' | sort -n | awk '{arr[NR]=$1} END {if (NR%2==1) print arr[(NR+1)/2]; else print (arr[NR/2]+arr[NR/2+1])/2}'
-
-  ##TO find MEAN
-  #grep -i -f files/genes/some_org/cat1.rna_list files/genes/some_org/cat1.gtf_slice | grep -i "utr" | grep -i "three\|3" | awk '{print $5-$4}' | awk '{ sum += $1; n++ } END { if (n > 0) print sum / n; }'
-
-  if [[ -s $TEMP_PATH/"$s_name"_"$reg.bed" ]]; then 
-    ln -r -f -s $TEMP_PATH/"$s_name"_"$reg.bed" $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed
-  elif [[ -s $TEMP_PATH/"$s_name"_cds.bed && $reg!="cds" ]]; then
-    if [[ "$reg" == "3utr" ]]; then
-      local flank_len=$(grep -w -i "$gene_full" files/genes/$org_name/gtf_stats.csv | awk -F',' '{ if ($9 > 3) sum += $9; n++ } END { if (n > 0) print sum / n; }') ##3' UTR length must be > 3 (because of existence of stop codons)
-      if [[ $flank_len == "" ]]; then
-        return 2
-      fi
-      bedtools flank -s -l $flank_len -r 0 -i $TEMP_PATH/"$s_name"_cds.bed -g $genome_fa.fai > $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed 
-    elif [[ "$reg" == "5utr" ]]; then
-      local flank_len=$(grep -w -i "$gene_full" files/genes/$org_name/gtf_stats.csv | awk -F',' '{ if ($8 > 0) sum += $8; n++ } END { if (n > 0) print sum / n; }') ##5' UTR length must be > 0
-      if [[ $flank_len == "" ]]; then
-        return 2
-      fi
-      bedtools flank -s -r $flank_len -l 0 -i $TEMP_PATH/"$s_name"_cds.bed -g $genome_fa.fai > $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed 
-    fi
-    if [[ -s $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed ]]; then
-      sed -i "s/cds/$(echo $reg)_FLANK/g" $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed
-    else
-      >&2 echo "$reg FLANKS not found for $s_name"
-      return 1
-    fi
-  fi
-
-  if [[ -s $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed || -L $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed ]]; then
-    if [[ $LABEL_FASTA ==  "TRUE" ]] ; then
-      bedtools getfasta -s -split -fi $genome_fa -bed $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed -nameOnly -fullHeader > "$FASTA_PATH/$s_name.$reg.tmp" #NOTUSING name+ because it also gives coordinates
-    else
-      bedtools getfasta -s -split -fi $genome_fa -bed $TEMP_PATH/"$s_name"_"$reg"_FETCH.bed -nameOnly -fullHeader > "$FASTA_PATH/$s_name.$reg" #NOTUSING name+ because it also gives coordinates
-    fi
-  fi
-fi
-
-return 0
-}
-
-####
-
 ##ENTRYPOINT
 
 #Check if scripts exist
@@ -96,12 +20,11 @@ fi
 
 GENOME_FILE=$1
 ANNO_FILE=$2
+GENE_LIST=$3
 f_org_name=$4
 org_name=$(echo $f_org_name | awk '{ gsub(/[[:punct:]]/, " ", $0); print $1" "$2; } ;')
 
-export -f get_fasta
-
-source fasta_functions.sh
+source functions.sh
 
 GENOMES_PATH=$(grep -i -w "genomes_path" parameters.txt | check_param) 
 ANNOS_PATH=$(grep -i -w "annos_path" parameters.txt | check_param) 
@@ -146,12 +69,13 @@ rm -f $genome_pipe
 mkfifo $genome_pipe
 
 gfile_name=$GENOME_FILE
-if [[ ${GENOME_FILE##*.} ==  "gz" ]] ; then
-  gfile_name=${GENOME_FILE%.*}
-    zcat -f $GENOME_FILE > $genome_pipe & 
-    cat < $genome_pipe > $gfile_name & 
-    genome_proc_id=$(echo $!)
+if [[ ${GENOME_FILE##*.} == "gz" ]] ; then
+    gfile_name=${GENOME_FILE%.*}
 fi
+zcat -f $GENOME_FILE > $genome_pipe & 
+cat < $genome_pipe > $gfile_name &
+#zcat -f $GENOME_FILE > $gfile_name &
+genome_proc_id=$(echo $!)
 
 if [[ $(echo $ANNO_FILE | grep -q -i "gtf") ]] ; then
   file_name=${ANNO_FILE%.*}
@@ -160,12 +84,17 @@ if [[ $(echo $ANNO_FILE | grep -q -i "gtf") ]] ; then
   ANNO_FILE=$ANNOS_PATH/"$5".gtf.gz
 fi
 
->&1 color_FG_BG_Bold $Black $BG_Yellow "0.1 Building Reference Genome Index (Samtools faidx)..."
+>&1 color_FG_BG_Bold $Black $BG_Yellow "Building Genome Index (Samtools faidx)..."
 
-if [[ ! -s $GENOME_FILE.fai ]] ; then
-  time samtools faidx --fai-idx $GENOME_FILE.fai $genome_pipe & 
+#if [[ ! -s $gfile_name.fai ]] ; then
+  #if [[ ${GENOME_FILE##*.} == "gz" ]] ; then
+  #  time samtools faidx -o $gfile_name --fai-idx $gfile_name.fai <(cat $genome_pipe) &
+  #else
+  #  time samtools faidx --fai-idx $gfile_name.fai <(cat $genome_pipe) &
+  #fi
+  time samtools faidx --fai-idx $gfile_name.fai <(cat $genome_pipe) &
   genome_index_proc=$(echo $!)
-fi
+#fi
 
 while [[ $(lsof -R -p $$ | wc -l) -gt $(($(ulimit -Sn)-200)) ]]; #limit file descriptors for child & parent process to 1024-200 
 do
@@ -175,7 +104,7 @@ done
 
 ###################################################################################################
 
-gene_list=($(cat $3 | sort | uniq | grep -v -w -i "gene"))
+gene_list=($(cat $GENE_LIST | sort | uniq | grep -v -w -i "gene"))
 
 >&1 color_FG_BG_Bold $Black $BG_Yellow "1. Checking Gene Names & Splitting GTFs for parallel processing..."
 
@@ -201,7 +130,7 @@ fi
 >&1 color_FG_BG_Bold $Black $BG_Yellow "2. Checking OrthoDB for missing & orthologous genes..."
 
 if [[ ! -s files/genes/$f_org_name/odb.list || ! -s files/genes/$f_org_name/final.list || ! -s files/genes/$f_org_name/odb.final_map ]] ; then
-  time ./check_OrthoDB.sh $f_org_name $4 files/genes/$f_org_name/odb.list $ANNO_FILE 
+  time ./check_OrthoDB.sh $f_org_name $GENE_LIST files/genes/$f_org_name/odb.list $ANNO_FILE 
 fi
 
 if [[ $(awk 'END{print NR;}' "files/genes/$f_org_name/odb.list" | awk '{print $1}') !=  0 ]] ; then
@@ -215,7 +144,7 @@ if [[ $(awk 'END{print NR;}' "files/genes/$f_org_name/odb.list" | awk '{print $1
   fi
 else
   >&2 echo $(color_FG_Bold $Red "2. Error : ")$(color_FG_BG_Bold $White $BG_Red "files/genes/$f_org_name/odb/odb.list")$(color_FG_Bold $Red " missing, Possibly orthologous genes were not found ")
-  >&2 color_FG_Bold $Red "2. If unsure, re-run command: ./jobhold.sh check_ortho ./check_OrthoDB.sh $f_org_name $4 $ANNO_FILE"
+  >&2 color_FG_Bold $Red "2. If unsure, re-run command: ./jobhold.sh check_ortho ./check_OrthoDB.sh $f_org_name $GENE_LIST $ANNO_FILE"
 fi
 
 ##REFRESH geene list based on file names
@@ -231,9 +160,9 @@ time Rscript --vanilla --verbose extract_gtf_info.R $TEMP_PATH/$f_org_name/ $f_o
 sed 1d files/genes/$f_org_name/gtf_stats.csv | awk -F',' '{print $1"\n"}' | sort | uniq | awk 'NF' > files/genes/$f_org_name/final.list
 
 if [[ ! -s files/genes/$f_org_name/gtf_stats.csv || ! -s files/genes/$f_org_name/final.list ]] ; then
-  >2& color_FG_Bold $Red "3. ERROR: Extraction of transcript stats failed..."
-  >2& color_FG_Bold $Red "3. Check $TEMP_PATH/$f_org_name/get_GTF_info.[o/e]"
-  >2& color_FG_Bold $Red "3. Remove $TEMP_PATH/$f_org_name/ & files/genes/$f_org_name/gtf_stats.csv and re-run the pipeline"
+  >&2 color_FG_Bold $Red "3. ERROR: Extraction of transcript stats failed..."
+  >&2 color_FG_Bold $Red "3. Check $TEMP_PATH/$f_org_name/get_GTF_info.[o/e]"
+  >&2 color_FG_Bold $Red "3. Remove $TEMP_PATH/$f_org_name/ & files/genes/$f_org_name/gtf_stats.csv and re-run the pipeline"
   exit 1
 fi
 
@@ -271,17 +200,17 @@ fi
 >&1 color_FG_BG_Bold $Black $BG_Yellow "5. Generating Metadata and Cleaning up..."
 
 sed 1d files/genes/$f_org_name/gtf_stats.csv | awk -F',' '{print $2}' | sort | uniq >  files/genes/$f_org_name/AVAILABLE_GENES
-grep -v -i -f files/genes/$f_org_name/AVAILABLE_GENES $4 | sort | uniq > files/genes/$f_org_name/MISSING_GENES
+grep -v -i -f files/genes/$f_org_name/AVAILABLE_GENES $GENE_LIST | sort | uniq > files/genes/$f_org_name/MISSING_GENES
 
 find $FASTA_PATH/$f_org_name/ -type f -name "*.fai" -exec rm -f {} +
 
-if [[ ! -z $gfile_name ]]; then
+if [[ ${GENOME_FILE##*.} == "gz" && ! -z $gfile_name ]]; then
   rm -f $gfile_name
 fi
 
 rm -rf $TEMP_PATH/$f_org_name/
 rm -f $bed_prefix/$f_org_name_*
-if [[ $REMOVE_DOWNLOADS ==  "TRUE" ]] ; then
+if [[ $REMOVE_DOWNLOADS == "TRUE" ]] ; then
    rm $GENOME_FILE
    rm $ANNO_FILE
 fi
